+++ lib/msal-browser/src/crypto/GuidGenerator.ts
