+++ lib/msal-browser/src/encode/Base64Decode.ts
