+++ lib/msal-browser/src/encode/Base64Encode.ts
